CALS-5551 change tel number type from int to long

+++ legacy-data-access/src/main/java/gov/ca/cwds/data/legacy/cms/entity/SubstituteCareProvider.java
@@ -354,7 +354,7 @@ public class SubstituteCareProvider implements PersistentObject {
     return addTelNo;
   }
 
-  public void setAddTelNo(int addTelNo) {
+  public void setAddTelNo(long addTelNo) {
     this.addTelNo = addTelNo;
   }
